Borré carpeta dts
